Modularize codebase into JS, CSS, and HTML components
- Created separate directories for JS, CSS, and potentially HTML.
- Moved existing script.js and style.css to their respective new directories (js/ and css/).
- Updated index.html to reflect new asset paths.
- Broke down the main script.js into smaller, more manageable modules stored in js/modules/.
  - Each module (Utils, TabManager, FeedManager, etc.) is now in its own file.
  - Modules use ES6 import/export syntax.
  - The main js/script.js now imports and initializes these modules.
- Added `type="module"` to the script tag in index.html.
- Refactored modules to check for DOM element existence before manipulation, improving robustness.
